pyproject.toml
[tool.poetry]
name = "todo-fastapi"
version = "0.1.0"
description = "Async FastAPI TODO list with DAO pattern, JWT auth, pagination, throttling"
authors = ["You <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
python-dotenv = "^1.0.0"
passlib = {extras=["bcrypt"], version="^1.7.4"}
pyjwt = "^2.8.0"
SQLAlchemy = {version = "^2.0.30", extras=["asyncio"]}
asyncpg = "^0.29.0"
email-validator = "^2.1.0"
pydantic = "^2.8.0"
slowapi = "^0.1.9"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"
black = "^24.4.0"
ruff = "^0.4.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
